fix usage of this.skip()
LDT and CDT Map tests should only run if the required feature is
available/enabled server side. With mocha.js v2 it was enough to call
this.skip() in during a test to abort it and mark it as pending. With
mocha.js v3 this behavior seems to have changed: After calling
this.skip() we also need to ensure the rest of the test case is not
executed any more as just calling this.skip() no longer aborts the test
directly.

resolves CLIENT-753